+++ src/main/kotlin/org/ornet/pat/localization/db/v2/LocalizedTextModel.kt
@@ -1,10 +1,10 @@
-package org.ornet.pat.localization.db.v1
+package org.ornet.pat.localization.db.v2
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class TomlModel(
+data class LocalizedTextModel(
     @SerialName("localizedText")
     val localizedTexts: List<LocalizedText>,
 )